rename package authentication
