+++ src/main/java/stanfordnlp/authentication/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package stanfordnlp.authentification;
+package stanfordnlp.authentication;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
